requirements.txt
sqlalchemy==2.0.43
asyncpg==0.30.0
fastapi==0.118.0
alembic==1.16.5
uvicorn==0.37.0
pydantic-settings==2.11.0
bcrypt==5.0.0
passlib==1.7.4
python-jose==3.5.0
greenlet==3.2.4